More refactoring, and begin making Daemon class work with asyncio
